requirements.txt
# FitCore Admin Panel - Backend Dependencies
# For Vercel + Supabase Deployment

# Core Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended>=4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1

# Database (PostgreSQL for Supabase)
SQLAlchemy>=2.0.23
psycopg2-binary>=2.9.0

# Security & Authentication
bcrypt==4.1.2
python-dotenv==1.0.0

# QR Code Generation (for QR Kiosk)
pyotp==2.9.0
qrcode==7.4.2
Pillow>=9.0.0

# Utilities
requests==2.31.0
python-dateutil>=2.8.2

# Production Server (for local testing, not used in Vercel)
gunicorn>=21.2.0
